+++ server/src/Room/room.service.ts
@@ -89,7 +89,13 @@ export class RoomService {
     return this.findOne(id);
   }
 
-  async deleteOne(id: number) {
-    return this.roomRepository.delete(id);
+  async deleteOne(id: number, userId: number) {
+    const foundRoom = await this.roomRepository.find({ id: id });
+    if (foundRoom[0].userId === userId) {
+      await this.roomRepository.delete(id);
+      return true;
+    } else {
+      return false;
+    }
   }
 }
